requirements.txt
discord.py[voice]
discord
yt-dlp
python-dotenv
PyNaCl